Set *out-of-repo* status string before tmgb string
Because the *out-of-repo* status string is set before, we keep using the
old syntax to set its style to remain compatible with older versions of
tmux-gitbar.

As the tmux-gitbar status string is set after and because we actually
have a specific style string for it, we can use the new syntax to set
the tmux-gitbar status string and style at the same time.

Note:

-old syntax:

```
set-window-option status-left-style STYLE
set-window-option status-left STATUS
```
-new syntax:

```
set-window-option status-left "#[STYLE]STATUS"
```